feat: introduce VolumeTimeSensitivityInterface and update imports for better organization
